+++ packages/scope-manager/src/lib/es2022.intl.ts
@@ -1,0 +1,11 @@
+// THIS CODE WAS AUTOMATICALLY GENERATED
+// DO NOT EDIT THIS CODE BY HAND
+// RUN THE FOLLOWING COMMAND FROM THE WORKSPACE ROOT TO REGENERATE:
+// npx nx generate-lib @typescript-eslint/scope-manager
+
+import { ImplicitLibVariableOptions } from '../variable';
+import { TYPE_VALUE } from './base-config';
+
+export const es2022_intl = {
+  Intl: TYPE_VALUE,
+} as Record<string, ImplicitLibVariableOptions>;